+++ spring/CHANGELOG.md
@@ -1,3 +1,11 @@
+## Next
+
+### Breaking Changes
+
+- rename `CreateBaseJsonResponseFunctions` to `CreateBaseJsonResponse`
+- rename `CreateSuccessJsonResponseFunctions` to `CreateSuccessJsonResponse`
+- rename `CreateFailureJsonResponseFunctions` to `CreateFailureJsonResponse`
+
 ## 0.1.0 (2026-03-16)
 
 initial release